# Improve round2 performance