@ReneWerner87 2020-11-19
```go
NEW
Benchmark_Round/round-12       1000000000     0.380 ns/op   0 B/op   0 allocs/op
Benchmark_Round/round-12       1000000000     0.439 ns/op   0 B/op   0 allocs/op
Benchmark_Round/round-12       1000000000     0.333 ns/op   0 B/op   0 allocs/op
Benchmark_Round/round-12       1000000000     0.310 ns/op   0 B/op   0 allocs/op

OLD
Benchmark_Round/round-12       126755714     9.24 ns/op   0 B/op   0 allocs/op
Benchmark_Round/round-12       124897803     9.32 ns/op   0 B/op   0 allocs/op
Benchmark_Round/round-12       117840432     9.64 ns/op   0 B/op   0 allocs/op
Benchmark_Round/round-12       133441935     9.66 ns/op   0 B/op   0 allocs/op
```

on my machine the difference is enormous, very good improvement