@ReneWerner87 2020-11-19
```go
NEW
Benchmark_Round-12	1000000000	 0.290 ns/op  0 B/op	   0 allocs/op
Benchmark_Round-12	1000000000	 0.292 ns/op  0 B/op	   0 allocs/op
Benchmark_Round-12	1000000000	 0.313 ns/op  0 B/op	   0 allocs/op
Benchmark_Round-12	1000000000	 0.276 ns/op  0 B/op	   0 allocs/op

OLD
Benchmark_Round-12	590925472	 2.46 ns/op	   0 B/op	   0 allocs/op
Benchmark_Round-12	512504625	 2.34 ns/op	   0 B/op	   0 allocs/op
Benchmark_Round-12	513369567	 2.42 ns/op	   0 B/op	   0 allocs/op
Benchmark_Round-12	494117672	 2.46 ns/op	   0 B/op	   0 allocs/op
```
@kiyonlin thx for the hint, this is the result with RunParallel