@kiyonlin 2020-11-28
Correct benchmark result, see golang/go#42865

```go
var result, input = 0, 2059

func BenchmarkRound2(b *testing.B) {
	var r int
	for i := 0; i < b.N; i++ {
		r = round2(input)
	}
	result = r
}
```

```hs
OLD
BenchmarkRound2-16      236888917                5.22 ns/op
BenchmarkRound2-16      240635983                5.11 ns/op
BenchmarkRound2-16      232815938                5.08 ns/op
BenchmarkRound2-16      238104736                5.07 ns/op
BenchmarkRound2-16      234706026                5.05 ns/op
BenchmarkRound2-16      237575134                5.11 ns/op
BenchmarkRound2-16      236359236                5.08 ns/op
BenchmarkRound2-16      235753626                5.11 ns/op
BenchmarkRound2-16      238996531                5.12 ns/op
BenchmarkRound2-16      236190168                5.03 ns/op

NEW
BenchmarkRound2-16      851576810                1.40 ns/op
BenchmarkRound2-16      834691317                1.37 ns/op
BenchmarkRound2-16      883740036                1.38 ns/op
BenchmarkRound2-16      863867149                1.38 ns/op
BenchmarkRound2-16      870876801                1.37 ns/op
BenchmarkRound2-16      846145068                1.42 ns/op
BenchmarkRound2-16      870322744                1.37 ns/op
BenchmarkRound2-16      874905960                1.36 ns/op
BenchmarkRound2-16      873266862                1.38 ns/op
BenchmarkRound2-16      873845012                1.37 ns/op
```

benchstat
```hs
name       old time/op  new time/op  delta
Round2-16  5.08ns ± 1%  1.37ns ± 1%  -73.01%  (p=0.000 n=9+8)
```

Sorry about the misleading @ReneWerner87